requirements.txt
PyMySQL==0.10.0
paho-mqtt==1.5.0
simplejson==3.13.2
h5py==2.10.0
Keras==2.3.1
librosa==0.7.2
matplotlib==3.1.2
numpy==1.16.4
pandas==1.1.0
Pillow==7.1.0
pydub==0.23.1
scikit-learn==0.22.2.post1